Populate requests.csv and meta.csv

+++ R/request.R
@@ -27,15 +27,62 @@ request <- function(solargis_dir, lat, lon, year, author, min_dist = 10000) {
 
     # TODO: Check the latitude value to see if it is bounded correctly.
 
-    requests_file <- paste(solargis_dir, "/requests.csv")
-    meta_file <- paste(solargis_dir, "/meta.csv")
+    requests_file <- paste0(solargis_dir, "/requests.csv")
+    meta_file <- paste0(solargis_dir, "/meta.csv")
 
     # Insert the request metadata into "requests.csv" for request history.
-    write.csv(data.frame(
-        lat = lat,
-        lon = lon,
-        year = year,
-        author = author,
-        min_dist = min_dist
-    ), file = requests_file, append = TRUE)
+    # Warnings are suppressed for when no file exists and a csv header must be
+    # created.
+    suppressWarnings(
+        write.table(data.frame(
+            lat = lat,
+            lon = lon,
+            year = year,
+            author = author,
+            min_dist = min_dist
+        ), file = requests_file, sep = ",", dec = ".", append = TRUE,
+        row.names = FALSE, col.names = !file.exists(requests_file))
+    )
+
+    # Find the closest existing request and see if the location is within the
+    # `min_dist` of the current requested location.
+    if (file.exists(meta_file)) {
+        meta <- data.table::fread(meta_file)
+
+        # Convert the two arrays of latitude and longitude to coordinate pairs.
+        coords <- mapply(c, meta$lat, meta$lon, SIMPLIFY = FALSE)
+
+        dist_to_prev_requests <- unlist(lapply(coords, function(coord) {
+            return(distance_between(lat, lon, coord[1], coord[2]))
+        }))
+
+        if (min(dist_to_prev_requests) < min_dist) {
+            message(paste("Found existing location within", min_dist, "meters."))
+
+            closest <- meta[which.min(dist_to_prev_requests), ]
+            file <- paste0(meta$file_hash, ".csv")
+            path <- paste(solargis_dir, "data", meta$year, file, sep = "/")
+
+            return(path)
+        }
+    } else {
+        message(paste("No existing location within", min_dist, "meters."))
+        message("Fetching new data set from SolarGIS...")
+
+        # Add the request to "meta.csv" because we are requesting a new dataset.
+        # Warnings are suppressed for a non-existing "meta.csv".
+        suppressWarnings(
+            write.csv(data.frame(
+                lat = lat,
+                lon = lon,
+                year = year,
+                file_hash = digest::sha1(paste(lat, lon, year, sep = "-"))
+            ), file = meta_file, sep = ",", dec = ".", append = TRUE,
+            row.names = FALSE, col.names = !file.exists(meta_file))
+        )
+
+        # TODO: Request data, write file & return path.
+
+        return(NULL)
+    }
 }
